[ci skip] Do not count deleted comments or streamable when processing user/streamable stats.

+++ src/main/java/com/graffitab/server/api/dto/streamable/FullStreamableDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 public class FullStreamableDto extends StreamableDto {
 
 	private Integer likersCount;
-	private Integer commentsCount;
+	private Long commentsCount;
 	private Boolean likedByCurrentUser;
 }
