+++ src/main/java/com/graffitab/server/api/dto/user/FullUserDto.java
@@ -9,5 +9,5 @@ public class FullUserDto extends UserDto {
 
 	private Integer followersCount;
 	private Integer followingCount;
-	private Integer streamablesCount;
+	private Long streamablesCount;
 }
